Move object mapping-related functions to a separate module
